requirements.txt
aniso8601==9.0.1
apturl==0.5.2
backcall==0.1.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
entrypoints==0.3
Flask==2.1.1
Flask-RESTful==0.3.9
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.3
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.15.2
Jinja2==3.1.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==2.1.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
parso==0.5.2
pbr==5.8.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prettytable==3.2.0
prompt-toolkit==2.0.10
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
systemd-python==234
tcping==0.1.1rc1
testresources==2.0.1
timeout-decorator==0.5.0
traitlets==4.3.3
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uWSGI==2.0.20
wadllib==1.3.3
wcwidth==0.1.8
Werkzeug==2.1.1
xkit==0.0.0
zipp==3.8.0
